added fav artist and album buttons

+++ api/controllers/favouriteController.js
@@ -1,12 +1,12 @@
 const ApiError = require('../error/ApiError');
-const { Favourite, FavouriteTrack } = require('../models/models');
+const { Favourite, FavouriteTrack, FavouriteArtist, FavouriteAlbum } = require('../models/models');
 
 
 class FavouriteController {
   async addOne(req, res, next) {
-    const { userId, trackId } = req.query;
+    const { userId, trackId, albumId, artistId } = req.query;
 
-    if (!userId || !trackId) {
+    if (!userId && !trackId && !albumId && !artistId) {
       return next(ApiError.badRequest('No id'));
     }
 
@@ -16,23 +16,56 @@ class FavouriteController {
       return next(ApiError.badRequest('favourite not found'));
     }
 
-    const candidate = await FavouriteTrack.findOne({where: {trackId,  favouriteId: favourite.id}})
+    if (trackId){
+      const candidate = await FavouriteTrack.findOne({where: {trackId,  favouriteId: favourite.id}})
 
-    if (candidate){
-      return next(ApiError.badRequest('Already in favourite'));
-    }
+      if (candidate){
+        return next(ApiError.badRequest('Already in favourite'));
+      }
 
-    const favouriteTrack = await FavouriteTrack.create({
+      const favouriteTrack = await FavouriteTrack.create({
       favouriteId: favourite.id,
       trackId
-    })
+      })
 
-    return res.json(favouriteTrack);
+      return res.json(favouriteTrack);
+    }
+
+    if (artistId){
+      const candidate = await FavouriteArtist.findOne({where: {artistId,  favouriteId: favourite.id}})
+
+      if (candidate){
+        return next(ApiError.badRequest('Already in favourite'));
+      }
+
+      const favouriteArtist = await FavouriteArtist.create({
+        favouriteId: favourite.id,
+        artistId
+      })
+
+      return res.json(favouriteArtist);
+    }
+
+    if (albumId){
+      const candidate = await FavouriteAlbum.findOne({where: {albumId,  favouriteId: favourite.id}})
+
+      if (candidate){
+        return next(ApiError.badRequest('Already in favourite'));
+      }
+
+      const favouriteAlbum = await FavouriteAlbum.create({
+        favouriteId: favourite.id,
+        albumId
+      })
+
+      return res.json(favouriteAlbum);
+    }
+
   }
   async removeOne(req, res, next) {
-    const { userId, trackId } = req.query;
+    const { userId, trackId, albumId, artistId } = req.query;
 
-    if (!userId || !trackId) {
+    if (!userId && !trackId && !albumId && !artistId) {
       return next(ApiError.badRequest('No id'));
     }
 
@@ -41,9 +74,20 @@ class FavouriteController {
     if (!favourite) {
       return next(ApiError.badRequest('favourite not found'));
     }
+    
+    let candidate
 
-    const candidate = await FavouriteTrack.destroy({where: {trackId,  favouriteId: favourite.id}})
+    if (trackId) {
+      candidate = await FavouriteTrack.destroy({where: {trackId,  favouriteId: favourite.id}})
+    }
 
+    if (albumId) {
+      candidate = await FavouriteAlbum.destroy({where: {albumId,  favouriteId: favourite.id}})
+    }
+
+    if (artistId) {
+      candidate = await FavouriteArtist.destroy({where: {artistId,  favouriteId: favourite.id}})
+    }
 
     return res.json(candidate);
   }
